Added HighScore Improved a little bit collision detection

+++ src/fr/cesi/ylalanne/game/WorldManager.java
@@ -13,71 +13,73 @@ public class WorldManager implements PropertyChangeListener {
 	private Player player;
 	private int playerRow;
 	private List<GameRow> rows;
-	private int playerX;
-	private int playerY;
+	private int lowestRow;
 	private ScheduledExecutorService moveExecutor;
+	private ScheduledExecutorService collisionExecutor;
 	
-	private void updateY(int y) {
-		playerY = y;
-		Optional<GameRow> currentRow = rows.stream()
-			.filter( gameRow -> gameRow.getBounds().isIn(y + (int)Math.ceil(player.getHeight() / 2.0d)) )
-			.findFirst();
+	
+	private void checkCollision() {
+		List<Obstacle> obstacles = world.getObstacles();
+		Optional<Obstacle> collider = obstacles.stream()
+		.filter(obstacle -> obstacle.isWithin(player.getX() + (int)Math.ceil(player.getWidth() / 2.0d), player.getY() + (int)Math.ceil(player.getHeight() / 2.0d), 25))
+		.filter( closeObstacle -> closeObstacle.checkCollision(player))
+		.findFirst();
 		
+		Optional<GameRow> currentRow = rows.stream()
+				.filter(row -> row.getBounds().isIn(player.getY() + (int)Math.ceil(player.getHeight()/2.0d)))
+				.findFirst();
 		if(currentRow.isPresent()) {
 			playerRow = rows.indexOf(currentRow.get()) +1;
-			int playerCenterX = player.getX() + (int)Math.ceil(player.getWidth() / 2.0d);
-			if(playerRow == 1 && currentRow.get().getWinningAreas().stream().anyMatch( area -> area.isIn(playerCenterX))) {
-				player.win();
-				world.end(true);
-			} else if (playerRow >= 2 && playerRow <= 4) {
-				long potentialColliders = world.getObstacles()
-					.stream()
-					.filter( obstacle -> obstacle.isWithin(playerX, y, 2))
-					.count();
-				
-				if(potentialColliders <= 0) {
-					player.kill();
-				}
+			if(playerRow < lowestRow) {
+				lowestRow = playerRow;
+				player.gains(10);
 			}
 		}
 		
-		checkPlayer();
-	}
-
-	private void checkPlayer() {
-		Optional<Obstacle> collider = world.getObstacles()
-				.stream()
-				.filter( obstacle -> obstacle.isWithin(playerX, playerY, 30) )
-				.filter( closeObstacle -> closeObstacle.checkCollision(player) )
-				.findFirst();
 		if(collider.isPresent()) {
 			player.collides(collider.get());
 		}
-	}
-
-	private void updateX(int x) {
-		playerX = x;
-		checkPlayer();
+		
+		int centerX = player.getX() + (int)Math.ceil(player.getWidth() / 2.0d);
+		int centerY = player.getY() + (int)Math.ceil(player.getHeight() / 2.0d);
+		
+		boolean hasVeryCloseObstacles = 
+				obstacles
+				.stream()
+				.anyMatch(obstacle -> obstacle.isWithin(centerX, centerY, 2));
+		
+		if(playerRow >=2 && playerRow <= 4 && !hasVeryCloseObstacles) {
+			player.kill();
+		} else if (playerRow == 1) {
+			int playerCenterX = player.getX() + (int)Math.ceil(player.getWidth() / 2.0d);
+			boolean won = rows.get(0).getWinningAreas()
+					.stream()
+					.anyMatch( winningRow -> winningRow.isIn(playerCenterX) );
+			if(won) {
+				player.win();
+				endWorld(true);
+			}
+		}
 	}
 
 	private void moveObstacles() {
 		for(Obstacle obstacle : world.getObstacles()) {
 			obstacle.move();
-			if(obstacle.isWithin(playerX, playerY, 30) && obstacle.checkCollision(player)) {
-				player.collides(obstacle);
-			}
 		}
 	}
-	
-	private void endWorld() {
+
+	private void endWorld(boolean won) {
 		try {
 			moveExecutor.shutdown();
 			moveExecutor.awaitTermination(500, TimeUnit.MILLISECONDS);
+			collisionExecutor.shutdown();
+			collisionExecutor.awaitTermination(500, TimeUnit.MILLISECONDS);
 		} catch (InterruptedException e) {
 			e.printStackTrace();
 			throw new RuntimeException(e);
 		} finally {
-			world.end(false);
+			player.removePropertyChangeListener(this);
+			world.end(won);
 		}
 	}
 
@@ -86,31 +88,25 @@ public class WorldManager implements PropertyChangeListener {
 		this.player = player;
 		this.playerRow = playerRow;
 		this.rows = rows;
-		playerX = player.getX();
-		playerY = player.getY();
 		this.player.addPropertyChangeListener(this);
+		lowestRow = rows.size();
 	}
 	
 	public void start() {
 		moveExecutor = Executors.newSingleThreadScheduledExecutor();
 		moveExecutor.scheduleAtFixedRate(this::moveObstacles, 0, 400, TimeUnit.MILLISECONDS);
+		collisionExecutor = Executors.newSingleThreadScheduledExecutor();
+		collisionExecutor.scheduleAtFixedRate(this::checkCollision, 0, 80, TimeUnit.MILLISECONDS);
 		player.lives();
 	}
 
 	@Override
 	public void propertyChange(PropertyChangeEvent evt) {
 		String propertyName = evt.getPropertyName();
-		Object newValue = evt.getNewValue();
 		
 		switch(propertyName) {
-			case "x":
-				updateX((int)newValue);
-				break;
-			case "y":
-				updateY((int)newValue);
-				break;
 			case "liveLess":
-				endWorld();
+				endWorld(false);
 				break;
 		}
